+++ src/Modmail.NET/Entities/GuildOption.cs
@@ -36,8 +36,6 @@ public sealed class GuildOption : IHasRegisterDate,
   public DateTime RegisterDateUtc { get; set; }
 
   public DateTime? UpdateDateUtc { get; set; }
-
-  public TeamPermissionLevel ManageGuildOptionMinAccessLevel { get; set; } = TeamPermissionLevel.Admin;
   public TeamPermissionLevel ManageTicketMinAccessLevel { get; set; } = TeamPermissionLevel.Moderator;
   public TeamPermissionLevel ManageTeamsMinAccessLevel { get; set; } = TeamPermissionLevel.Admin;
   public TeamPermissionLevel ManageBlacklistMinAccessLevel { get; set; } = TeamPermissionLevel.Admin;